src/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
python-multipart==0.0.20
python-dotenv===1.1.1
pydantic-settings==2.11.0
aiofiles==24.1.0
langchain==0.3.27
PyMuPDF==1.26.5
motor==3.7.1
openai==2.6.0
cohere==5.19.0
qdrant-client==1.15.1